chore(eigen_fee_metrics): stub out the eigen fee metrics and add to eigen committer
